+++ Check.ps1
@@ -1,0 +1,4 @@
+flutter format . 
+flutter doctor 
+flutter analyze
+flutter test