+++ lib/src/controllers/controller_manager.dart
@@ -1,13 +1,14 @@
 import '../../qlevar_router.dart';
 import '../routes/qroute_children.dart';
+import '../routes/qroute_internal.dart';
 import '../types/qroute_key.dart';
 import 'qrouter_controller.dart';
 
 class ControllerManager {
   final controllers = <QRouterController>[];
 
-  QRouterController createController(
-      String name, List<QRoute> routes, String? initPath) {
+  QRouterController createController(String name, List<QRoute> routes,
+      String? initPath, QRouteInternal? initRoute) {
     if (hasController(name)) {
       QR.log('A navigator with name [$name] already exist', isDebug: true);
       return controllers.firstWhere((element) => element.key.hasName(name));
@@ -17,10 +18,9 @@ class ControllerManager {
       throw Exception('Route with name $name was not found in the tree info');
     }
     final key = QKey(name);
-    final controller = QRouterController(
-        key,
+    final controller = QRouterController(key,
         QRouteChildren.from(routes, key, routePath == '/' ? '' : routePath),
-        initPath ?? '/');
+        initPath: initPath ?? '/', initRoute: initRoute);
     controllers.add(controller);
     return controller;
   }