+++ lib/src/qr.dart
@@ -2,6 +2,7 @@ import 'package:flutter/material.dart';
 
 import '../qlevar_router.dart';
 import 'controllers/controller_manager.dart';
+import 'controllers/match_controller.dart';
 import 'controllers/qrouter_controller.dart';
 import 'helpers/platform/configure_web.dart'
     if (dart.library.io) 'helpers/platform/configure_nonweb.dart';
@@ -51,9 +52,11 @@ class QRContext {
   QNavigator navigatorOf(String name) => _manager.withName(name);
 
   ///  return a router [QRouter] for the given routes
+  /// you do not need to give the [initRoute]
   QRouter createNavigator(String name, List<QRoute> routes,
-      {String? initPath}) {
-    final controller = createRouterController(name, routes, initPath: initPath);
+      {String? initPath, QRouteInternal? initRoute}) {
+    final controller = createRouterController(name, routes,
+        initPath: initPath, initRoute: initRoute);
     return QRouter(controller);
   }
 
@@ -81,8 +84,8 @@ class QRContext {
           addHistory: addHistory);
 
   /// Add this routes as child for the route with name.
-  //void expandRoute(String name, List<QRoute> routes) {}
-  /// Remove this route from the router
+  // void expandRoute(String name, List<QRoute> routes) {}
+  // Remove this route from the router
   //void cleanRoute(String routerName, String routeName) {}
 
   /// return the current tree widget
@@ -92,19 +95,29 @@ class QRContext {
 
   /// create a controller to use with a Navigator
   QRouterController createRouterController(String name, List<QRoute> routes,
-          {String? initPath}) =>
-      _manager.createController(name, routes, initPath);
+          {String? initPath, QRouteInternal? initRoute}) =>
+      _manager.createController(name, routes, initPath, initRoute);
 
   /// Navigate to this path.
   /// The package will try to get the right navigtor to this path.
-  Future<void> to(String path) async {
+  Future<void> to(String path, {bool ignoreSamePath = true}) async {
+    if (ignoreSamePath && currentPath == path) {
+      return;
+    }
     final controller = _manager.withName(rootRouterName);
     var match = controller.findPath(path);
     await _toMatch(match);
   }
 
   /// Go to a route with given name
-  Future<void> toName(String name, {Map<String, dynamic>? params}) async {
+  Future<void> toName(String name,
+      {Map<String, dynamic>? params, bool ignoreSamePath = true}) async {
+    if (ignoreSamePath &&
+        currentPath ==
+            MatchController.findPathFromName(
+                name, params ?? <String, dynamic>{})) {
+      return;
+    }
     final controller = _manager.withName(rootRouterName);
     var match = controller.findName(name, params: params);
     await _toMatch(match);
@@ -113,6 +126,10 @@ class QRContext {
   Future<void> _toMatch(QRouteInternal match,
       {String forController = QRContext.rootRouterName}) async {
     final controller = _manager.withName(forController);
+    // if (controller.isDeclarative) {
+    //   controller.updateDeclarative(match: match);
+    //   return;
+    // }
     await controller.popUnitOrPushMatch(match, checkChild: false);
     if (match.hasChild) {
       final newControllerName =