+++ lib/src/routers/qrouter.dart
@@ -6,6 +6,10 @@ class QRouter extends StatefulWidget {
   /// the Navigation key for this [Navigator]
   final navKey = GlobalKey<NavigatorState>();
 
+  final QRouterController _controller;
+
+  QRouter(this._controller);
+
   /// Get the name for the current child
   /// This is the name whih define in [QRoute.name] if it is null [QRoute.path]
   /// Will be returned
@@ -15,9 +19,6 @@ class QRouter extends StatefulWidget {
   /// Get the [QNavigator] for this [QRouter] to add or remove pages to it.
   QNavigator get navigator => _controller;
 
-  final QRouterController _controller;
-
-  QRouter(this._controller);
   @override
   _QRouterState createState() => _QRouterState();
 }
@@ -30,18 +31,28 @@ class _QRouterState extends State<QRouter> {
   }
 
   @override
-  Widget build(BuildContext context) => Navigator(
-        key: widget.navKey,
-        pages: widget._controller.pages,
-        onPopPage: (route, result) {
-          if (!route.didPop(result)) {
-            return false;
-          }
-          if (widget._controller.canPop) {
-            widget._controller.removeLast();
-            return true;
-          }
+  void didChangeDependencies() {
+    widget._controller.updateDeclarative();
+    super.didChangeDependencies();
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Navigator(
+      key: widget.navKey,
+      pages: widget._controller.pages,
+      onPopPage: (route, result) {
+        if (!route.didPop(result)) {
           return false;
-        },
-      );
+        }
+        return widget._controller.removeLast();
+      },
+    );
+  }
+
+  @override
+  void dispose() {
+    widget._controller.dispose();
+    super.dispose();
+  }
 }